Test EUCTR results scraper
This is a notebook that includes some testing for building a crawler that will scrape data from EUCTR results pages. Each step is full documented. A separate file will remove the testing and condense the code down to only what is necessary to run the crawlers. Note, this has only run limited testing for a few trial pages. It has not been tested on the full EUCTR yet.