A kinder gentler IRCT scraper
This used to brute force it's way through about twice as many urls as necessary in a brute force way. Now it pre-gathers all the url suffixes before it does the main scrape which is much more efficient.

I've also seperated the normal script from the one I have running with multiprocessing on mac.